backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
python-docx